linker information: Move into separate file
We use this struct in the future to read the eh frame address and size.
Therefore, we need it from different places.
